Win32: cleaned up most of C4244 warnings with 32-bit compilation.
C4244 warnings (conversion from 'type1' to 'type2', possible loss of data)
were disabled in 6861:e4590dfd97ff as part of introducing support for
64-bit compilation with MSVC.  Still, these warnings remain useful
in some cases.  Additionally, they explain multiple otherwise unneeded type
casts in the code.

This change resolves most of the warnings as observed during 32-bit
compilation with C4244 warnings enabled by introducing additional type
casts, in most cases already used in other parts of the code in similar
constructs.

+++ src/http/ngx_http_file_cache.c
@@ -387,7 +387,8 @@ ngx_http_file_cache_open(ngx_http_request_t *r)
 
     if (of.is_directio) {
 
-        c->body_start = ngx_align(c->body_start, clcf->directio_alignment);
+        c->body_start = ngx_align(c->body_start,
+                                  (size_t) clcf->directio_alignment);
 
         c->buf = ngx_calloc_buf(r->pool);
         if (c->buf == NULL) {
@@ -395,7 +396,7 @@ ngx_http_file_cache_open(ngx_http_request_t *r)
         }
 
         c->buf->start = ngx_pmemalign(r->pool, c->body_start,
-                                      clcf->directio_alignment);
+                                      (size_t) clcf->directio_alignment);
         if (c->buf->start == NULL) {
             return NGX_ERROR;
         }
